requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
automium==0.2.6
automium-web==0.1.1
backports.zoneinfo==0.2.1
blinker==1.7.0
click==8.1.7
Django==4.2.11
django-admin-sortable==2.3
django-admin-sortable2==2.1.10
django-ninja==1.1.0
djangorestframework==3.15.0
Flask==3.0.2
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
legacy==0.1.7
MarkupSafe==2.1.5
pillow==10.2.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
sqlparse==0.4.4
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.18.1
